requirements.txt
domaintools-api>=2.0.0
rich>=13.0.0
click>=8.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
aiohttp>=3.9.0
typing-extensions>=4.8.0
pyyaml>=6.0.0
